stm32: refactor wait logic into wait mod
