+++ embassy-stm32/src/adc/f3.rs
@@ -3,10 +3,10 @@ use core::marker::PhantomData;
 use stm32_metapac::adc::regs::{Smpr1, Smpr2, Sqr1, Sqr2, Sqr3};
 use stm32_metapac::adc::vals::{Advregen, Dmacfg};
 
-use super::blocking_delay_us;
 use crate::adc::{Adc, AdcRegs, ConversionMode, DefaultInstance, Instance, SampleTime, VrefInt};
 use crate::interrupt::typelevel::Interrupt;
 use crate::time::Hertz;
+use crate::wait::block_for_us;
 use crate::{Peri, interrupt, rcc};
 
 pub const VDDA_CALIB_MV: u32 = 3300;
@@ -154,7 +154,7 @@ impl<'d, T: DefaultInstance> Adc<'d, T> {
         T::regs().cr().modify(|w| w.set_advregen(Advregen::Enabled));
 
         // Wait for the regulator to stabilize
-        blocking_delay_us(10);
+        block_for_us(10);
 
         assert!(!T::regs().cr().read().aden());
 
@@ -165,7 +165,7 @@ impl<'d, T: DefaultInstance> Adc<'d, T> {
         while T::regs().cr().read().adcal() {}
 
         // Wait more than 4 clock cycles after adcal is cleared (RM0364 p. 223).
-        blocking_delay_us((1_000_000 * 4) / Self::freq().0 as u64 + 1);
+        block_for_us((1_000_000 * 4) / Self::freq().0 as u64 + 1);
 
         // Enable the adc
         T::regs().enable();
