+++ embassy-stm32/src/adc/v2.rs
@@ -1,10 +1,13 @@
 use core::sync::atomic::{Ordering, compiler_fence};
 
-use super::{AnyAdcChannel, ConversionMode, Temperature, Vbat, VrefInt, blocking_delay_us};
-use crate::adc::{Adc, AdcRegs, DefaultInstance, InjectedRegs, Resolution, SampleTime};
+use crate::adc::{
+    Adc, AdcRegs, AnyAdcChannel, ConversionMode, DefaultInstance, InjectedRegs, Resolution, SampleTime, Temperature,
+    Vbat, VrefInt,
+};
 use crate::pac::adc::vals;
 pub use crate::pac::adccommon::vals::Adcpre;
 use crate::time::Hertz;
+use crate::wait::block_for_us;
 use crate::{Peri, rcc};
 
 mod injected;
@@ -92,7 +95,7 @@ impl AdcRegs for crate::pac::adc::Adc {
             reg.set_adon(true);
         });
 
-        blocking_delay_us(3);
+        block_for_us(3);
     }
 
     fn start(&self) {
