+++ embassy-stm32/src/dsihost/mod.rs
@@ -14,7 +14,8 @@ use crate::peripherals::{self};
 use crate::rcc::dsi::DSI_CONFIG;
 use crate::rcc::{self, RccPeripheral};
 use crate::time::MaybeHertz;
-use crate::{Peri, block_for_us, interrupt};
+use crate::wait::block_for_us;
+use crate::{Peri, interrupt};
 
 mod phy;
 pub use phy::{DsiHostPhyConfig, DsiHostPhyLanes};
