+++ embassy-stm32/src/wait.rs
@@ -1,8 +1,26 @@
 //! Provides a function to try until something is true
 
+use crate::rcc;
+
+/// Performs a busy-wait delay for a specified number of microseconds that is async if possible
+#[allow(dead_code)]
+pub async fn wait_for_us(us: u64) {
+    #[cfg(feature = "time")]
+    embassy_time::Timer::after_micros(us).await;
+
+    #[cfg(not(feature = "time"))]
+    block_for_us(us);
+}
+
+/// Performs a busy-wait delay for a specified number of microseconds.
+#[allow(dead_code)]
+pub fn block_for_us(us: u64) {
+    cortex_m::asm::delay(unsafe { rcc::get_freqs().sys.to_hertz().unwrap().0 as u64 * us / 1_000_000 } as u32);
+}
+
 #[cfg(feature = "time")]
 /// Function to try until something is true
-#[allow(unused)]
+#[allow(dead_code)]
 pub async fn try_until(mut func: impl AsyncFnMut() -> bool, micros: u64) -> Result<(), ()> {
     use embassy_time::{Duration, Ticker};
 
@@ -24,7 +42,7 @@ pub async fn try_until(mut func: impl AsyncFnMut() -> bool, micros: u64) -> Resu
 
 #[cfg(not(feature = "time"))]
 /// Function to try until something is true
-#[allow(unused)]
+#[allow(dead_code)]
 pub async fn try_until(mut func: impl AsyncFnMut() -> bool, micros: u64) -> Result<(), ()> {
     use embassy_futures::yield_now;
 
@@ -35,7 +53,7 @@ pub async fn try_until(mut func: impl AsyncFnMut() -> bool, micros: u64) -> Resu
             return Ok(());
         }
 
-        crate::block_for_us(1_000);
+        block_for_us(1_000);
         yield_now().await;
     }
